+++ radical_translations/core/urls.py
@@ -5,6 +5,7 @@ from radical_translations.core.views import (
     ResourceDetailView,
     ResourceViewSet,
     SimpleResourceViewSet,
+    network,
     resource_list,
 )
 
@@ -17,4 +18,5 @@ router.register(
 urlpatterns = [
     path("<int:pk>/", ResourceDetailView.as_view(), name="resource-detail"),
     path("", resource_list, name="resource-list"),
+    path("network/", network, name="resource-network"),
 ] + router.urls