+++ CHANGELOG.rst
@@ -20,6 +20,8 @@ Added
 * Zoom feature to the timeline grid.
 * Sorting options to the base settings.
 * Wikidata as an option for `Agent` page.
+* `Resources` network visualisation.
+* `Agents` network visualisation.
 
 Changed
 ~~~~~~~